+++ .lintstagedrc.mjs
@@ -18,12 +18,10 @@ export default {
   "{packages,examples-extra/basic}/**/*.{js,jsx,ts,tsx,mjs,cjs}": (
     files,
   ) => {
-    console.error(files);
     const match = micromatch.not(
       files,
       ["**/templates/**/*", "**/generatedNoCheck/**/*"],
     );
-    console.error(match.join(" "));
     if (match.length === 0) return [];
     return [
       `dprint fmt ${match.join(" ")}`,