+++ packages/client/src/ontology/StandardOntologyProvider.ts
@@ -29,7 +29,7 @@ import type {
 } from "@osdk/internal.foundry";
 import deepEqual from "fast-deep-equal";
 import type { MinimalClient } from "../MinimalClientContext.js";
-import { createAsyncCache } from "../object/Cache.js";
+import { createAsyncClientCache } from "../object/Cache.js";
 import { loadFullObjectMetadata } from "./loadFullObjectMetadata.js";
 import { loadInterfaceDefinition } from "./loadInterfaceDefinition.js";
 import type { OntologyProviderFactory } from "./OntologyProvider.js";
@@ -91,7 +91,8 @@ export const createStandardOntologyProviderFactory: (
             true,
           ),
           implements: fullCacheLocal[0].data.map((i) => i.apiName),
-        } as ObjectTypeDefinition<any, any>;
+          rid: fullCacheLocal[1].objectTypes[key].objectType.rid,
+        } as ObjectTypeDefinition<any, any> & { rid: string };
       } else {
         return await loadFullObjectMetadata(client, key);
       }
@@ -121,7 +122,7 @@ export const createStandardOntologyProviderFactory: (
         skipCache?: boolean,
       ) => Promise<N>,
     ) {
-      const cache = createAsyncCache<string, N>((client, key) =>
+      const cache = createAsyncClientCache<string, N>((client, key) =>
         fn(client, key, false)
       );
       return async (apiName: string) => {