+++ packages/client/src/ontology/makeConjureContext.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Palantir Technologies, Inc. All rights reserved.
+ * Copyright 2024 Palantir Technologies, Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,17 @@
  * limitations under the License.
  */
 
-export { UNSTABLE_createClient as createClient } from "../__unstable/UNSTABLE_createClient.js";
-export type { UnstableClient as Client } from "../__unstable/UnstableClient.js";
-export { augment } from "../object/fetchPage.js";
-export type { UNSTABLE_ObjectSet as ObjectSet } from "../objectSet/createUnstableObjectSet.js";
-export type { ObjectSetListener } from "../objectSet/ObjectSetListener.js";
+import type { SharedClientContext } from "@osdk/shared.net";
+import type { ConjureContext } from "conjure-lite";
+
+export function makeConjureContext(
+  { baseUrl, fetch: fetchFn, tokenProvider }: SharedClientContext,
+  servicePath: string,
+): ConjureContext {
+  return {
+    baseUrl,
+    servicePath,
+    fetchFn,
+    tokenProvider,
+  };
+}