+++ packages/foundry.core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/foundry.core
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [0a64def]
+- Updated dependencies [f9b3c72]
+- Updated dependencies [978ecd5]
+- Updated dependencies [978ecd5]
+- Updated dependencies [c9f3214]
+  - @osdk/client@0.19.0
+  - @osdk/api@1.8.0
+  - @osdk/shared.net@1.11.0
+
 ## 0.1.0
 
 ### Minor Changes