+++ packages/generator-converters/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @osdk/generator-converters
 
+## 0.6.0
+
+### Patch Changes
+
+- Updated dependencies [c9f3214]
+- Updated dependencies [c9f3214]
+  - @osdk/gateway@2.3.0
+  - @osdk/api@1.8.0
+
 ## 0.5.0
 
 ### Patch Changes